w3c-reports/12092014/gtk/runned-tests/user-timing/test_user_timing_measure_navigation_timing-expected.txt

Description

This test validates that the performance.measure() method is working properly when navigation timing attributes are used in place of mark names. This test creates the following measures to test this method:

"measure_nav_start_no_end": created using a measure() call with a navigation timing attribute provided as the startMark and nothing provided as the endMark
"measure_nav_start_mark_end": created using a measure() call with a navigation timing attribute provided as the startMark and a mark name provided as the endMark
"measure_mark_start_nav_end": created using a measure() call with a mark name provided as the startMark and a navigation timing attribute provided as the endMark
"measure_nav_start_nav_end":created using a measure() call with a navigation timing attribute provided as both the startMark and endMark
After creating each measure, the existence of these measures is validated by calling performance.getEntriesByName() with each measure name

PASS window.performance is defined 
FAIL The User Timing and Performance Timeline interfaces, which are required for this test, are defined. assert_true: The User Timing and Performance Timeline interfaces, which are required for this test, are defined. expected true got false(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_true@http://localhost:8800/resources/testharness.js:299:15
http://localhost:8800/user-timing/resources/webperftestharness.js:134:38
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
wp_test@http://localhost:8800/user-timing/resources/webperftestharness.js:62:9
test_true@http://localhost:8800/user-timing/resources/webperftestharness.js:134:12
onload_test@http://localhost:8800/user-timing/test_user_timing_measure_navigation_timing.html:107:26
onload@http://localhost:8800/user-timing/test_user_timing_measure_navigation_timing.html:213:12)
